Extract logging out of repository_ext.rs
