+++ crates/gitbutler-branch-actions/src/base.rs
@@ -14,7 +14,10 @@ use gitbutler_error::error::Marker;
 use gitbutler_oxidize::{git2_to_gix_object_id, gix_to_git2_oid, GixRepositoryExt};
 use gitbutler_project::FetchResult;
 use gitbutler_reference::{Refname, RemoteRefname};
-use gitbutler_repo::{LogUntil, RepositoryExt};
+use gitbutler_repo::{
+    logging::{LogUntil, RepositoryExt as _},
+    RepositoryExt,
+};
 use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::{BranchOwnershipClaims, Stack, Target, VirtualBranchesHandle};
 use serde::Serialize;
