+++ crates/gitbutler-branch-actions/src/integration.rs
@@ -11,8 +11,9 @@ use gitbutler_error::error::Marker;
 use gitbutler_operating_modes::OPEN_WORKSPACE_REFS;
 use gitbutler_oxidize::{git2_to_gix_object_id, gix_to_git2_oid, GixRepositoryExt};
 use gitbutler_project::access::WorktreeWritePermission;
+use gitbutler_repo::logging::{LogUntil, RepositoryExt as _};
+use gitbutler_repo::RepositoryExt;
 use gitbutler_repo::SignaturePurpose;
-use gitbutler_repo::{LogUntil, RepositoryExt};
 use gitbutler_stack::{Stack, VirtualBranchesHandle};
 use tracing::instrument;
