+++ crates/gitbutler-branch-actions/src/undo_commit.rs
@@ -4,7 +4,10 @@ use anyhow::{bail, Context as _, Result};
 use gitbutler_command_context::CommandContext;
 use gitbutler_commit::commit_ext::CommitExt as _;
 use gitbutler_diff::Hunk;
-use gitbutler_repo::{rebase::cherry_rebase_group, LogUntil, RepositoryExt as _};
+use gitbutler_repo::{
+    logging::{LogUntil, RepositoryExt as _},
+    rebase::cherry_rebase_group,
+};
 use gitbutler_stack::{OwnershipClaim, Stack, StackId};
 
 use crate::VirtualBranchesExt as _;
