+++ crates/gitbutler-branch-actions/src/virtual.rs
@@ -28,8 +28,9 @@ use gitbutler_oxidize::{
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_reference::{normalize_branch_name, Refname, RemoteRefname};
 use gitbutler_repo::{
+    logging::{LogUntil, RepositoryExt as _},
     rebase::{cherry_rebase, cherry_rebase_group},
-    LogUntil, RepositoryExt,
+    RepositoryExt,
 };
 use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::{
